+++ GestaoTalentos.API/AuthDtos.cs
@@ -25,4 +25,14 @@ public record RecordUpdateDto(string Content, bool IsShared);
 public record ClienteDto(int Id, string Nome, string Email, int IdCriador, int? IdMinhaConta);
 
 /// DTO para criação de um novo cliente.
-public record ClienteCreateDto(string Nome, string Email, int? IdMinhaConta);
+public record ClienteCreateDto(string Nome, string Email, int? IdCriador, int? IdMinhaConta);
+/// DTO para atualização de um cliente existente.
+public record ClienteUpdateDto(string Nome, string Email, int? IdCriador, int? IdMinhaConta);
+
+/// DTO para representação de uma apresentação na resposta da API.
+//public record ApresentacaoDto(int IdCliente, DateTime DataApresentacao);
+
+/// DTO para criação de uma nova apresentação.
+//public record ApresentacaoCreateDto(int IdCliente, DateTime DataApresentacao);
+/// DTO para atualização de uma apresentação existente.
+//public record ApresentacaoUpdateDto(int IdCliente, ;
