+++ GestaoTalentos.Domain/GestaoTalentos.Domain.csproj
@@ -7,3 +7,5 @@
   </PropertyGroup>
 
 </Project>
+
+
