+++ GestaoTalentos.Domain/IApresentacaoRepository.cs
@@ -1,0 +1,38 @@
+//namespace GestaoTalentos.Domain;
+
+/// Interface para repositório de clientes, definindo operações CRUD básicas.
+//public interface IApresentacaoRepository
+//{
+
+    /// Obtém uma Apresentacao pelo Id do Cliente e Id do Talento
+    //Task<Apresentacao?> GetByIdsAsync(int idCliente, int idTalento);
+    
+    //Task<Apresentacao?> GetByIdClientesync(int idCliente);
+    
+    
+    /// Obtém todas as Apresentações.
+    /// <returns>Lista de todos as Apresentações.</returns>
+    //  Task<List<Apresentacao>> GetAllAsync();
+    
+    /// Adiciona uma nova Apresentacao.
+    /// <param name="apresentacao">Instância do Apresentacao a ser adicionada.</param>
+    //Task AddAsync(Apresentacao apresentacao);
+    
+    /// Atualiza uma apresentacao existente.
+    /// <param name="apresentacao">Instância do Apresentacao a ser atualizada.</param>
+    //Task UpdateAsync(Apresentacao apresentacao);
+    
+    /// Verifica se existe pelo menos uma apresentacao.
+    /// <returns>True se existir pelo menos uma apresentacao, false caso contrário.</returns>
+    //Task<bool> AnyAsync();
+    
+    /// Lista todas as apresentações associadas a um cliente específico, identificando-o pelo seu ID.
+    /// <param name="idCliente"></param>
+    /// <returns></returns>
+    //Task<List<Apresentacao>> GetByClienteIdAsync(int idCliente);
+
+    /// Lista todas as apresentações associadas a um talento específico, identificando-o pelo seu ID.
+    /// <param name="idTalento"></param>
+    /// <returns></returns>
+    //Task<List<Apresentacao>> GetByTalentoIdAsync(int idTalento);
+//}
